// public/src/components/ui/elements/constants/enum.ts
export enum ElementType  {
    Input = 'Input',
    InputNumber = 'InputNumber',
    Select = 'Select',
    RadioGroup = 'RadioGroup',
    Textarea = 'Textarea',
    InputCompact = 'InputCompact',
    Switch = 'Switch',
    Checkbox = 'Checkbox',
    UploadFile = 'UploadFile',
    CodeEditor = 'CodeEditor',
    Collapse = 'Collapse',
    // Button = 'Button',
    // Link = 'Link',
    // Image = 'Image',
    
   
}